[0.99] Move Config to Lookup and document
* Configuring using `Timezone::Lookup::config` makes a lot more sense
  than `Timezone::Config::config` because we are only configuring
  lookups.
* Added documentation to the class.
